made DummyTestTriple test class, refactored test file structure
