+++ simple/src/test/java/org/apache/commons/rdf/simple/DatasetGraphViewTest.java
@@ -1,6 +1,10 @@
 package org.apache.commons.rdf.simple;
 
-import org.apache.commons.rdf.simple.experimental.DummyTestDataset;
+import org.apache.commons.rdf.api.BlankNodeOrIRI;
+import org.apache.commons.rdf.api.IRI;
+import org.apache.commons.rdf.api.RDFTerm;
+import org.apache.commons.rdf.api.Triple;
+import org.apache.commons.rdf.simple.dummy.DummyTestDataset;
 import org.junit.Test;
 
 /**
@@ -18,5 +22,41 @@ public class DatasetGraphViewTest {
         } catch (Exception e) {
             e.printStackTrace();
         }
+    }
+
+    @Test
+    public void addTest(){
+
+            datasetGraphView.add(new Triple() {
+                @Override
+                public BlankNodeOrIRI getSubject() {
+                    return new BlankNodeImpl();
+                }
+
+                @Override
+                public IRI getPredicate() {
+                    return new IRI() {
+                        @Override
+                        public String getIRIString() {
+                            return "getIRIString";
+                        }
+
+                        @Override
+                        public String ntriplesString() {
+                            return "ntriplesString";
+                        }
+                    };
+                }
+
+                @Override
+                public RDFTerm getObject() {
+                    return new RDFTerm() {
+                        @Override
+                        public String ntriplesString() {
+                            return "ntriplesString";
+                        }
+                    };
+                }
+            });
     }
 }
